feat: separate logic into different files
