+++ src/lib.rs
@@ -1,4 +1,7 @@
-pub mod sn_membership;
+pub mod consensus;
+pub mod membership;
+pub mod proposal;
+pub mod vote;
 
 // #[cfg(feature = "bad_crypto")]
 // pub mod bad_crypto;
@@ -7,14 +10,19 @@ pub mod sn_membership;
 // #[cfg(feature = "ed25519")]
 // pub mod ed25519;
 
-pub use crate::sn_membership::{Ballot, Generation, Reconfig, SignedVote, State, Vote};
-
 // #[cfg(feature = "bad_crypto")]
 // pub use crate::bad_crypto::{PublicKey, SecretKey, Signature};
 // #[cfg(feature = "blsttc")]
 // pub use crate::blsttc::{PublicKey, SecretKey, Signature};
 // #[cfg(feature = "ed25519")]
 // pub use crate::ed25519::{PublicKey, SecretKey, Signature};
+
+pub use blsttc::{PublicKeyShare, SecretKeyShare, SignatureShare};
+
+pub use crate::consensus::State;
+pub use crate::membership::{Generation, MembershipState, Reconfig};
+pub use crate::proposal::Proposal;
+pub use crate::vote::{Ballot, Proposition, UnsignedVote, Vote};
 
 pub mod error;
 pub use crate::error::Error;
