+++ src/proposal.rs
@@ -1,0 +1,5 @@
+use crate::Result;
+
+pub trait Proposal {
+    fn validate(&self) -> Result<()>;
+}
