+++ tests/sn_membership.rs
@@ -14,13 +14,15 @@ mod net;
 
 use quickcheck::{Arbitrary, Gen, TestResult};
 use quickcheck_macros::quickcheck;
-use sn_membership::{Ballot, Error, Generation, Reconfig, Result, SignedVote, State, Vote};
+use sn_membership::{
+    Ballot, Error, Generation, MembershipState, Reconfig, Result, UnsignedVote, Vote,
+};
 
 #[test]
 fn test_reject_changing_reconfig_when_one_is_in_progress() -> Result<()> {
     let mut rng = StdRng::from_seed([0u8; 32]);
-    let mut proc: State<u8> = State::random(&mut rng);
-    proc.elders = BTreeSet::from_iter([proc.public_key_share()]);
+    let mut proc: MembershipState<u8> = MembershipState::random(&mut rng);
+    proc.state.elders = BTreeSet::from_iter([proc.public_key_share()]);
     proc.propose(Reconfig::Join(rng.gen()))?;
     assert!(matches!(
         proc.propose(Reconfig::Join(rng.gen())),
@@ -35,8 +37,8 @@ fn test_reject_vote_from_non_member() -> Result<()> {
     let mut net = Net::with_procs(2, &mut rng);
     let p0 = net.procs[0].public_key_share();
     let p1 = net.procs[1].public_key_share();
-    net.procs[0].elders = BTreeSet::from_iter([p0]);
-    net.procs[1].elders = BTreeSet::from_iter([p0, p1]);
+    net.procs[0].state.elders = BTreeSet::from_iter([p0]);
+    net.procs[1].state.elders = BTreeSet::from_iter([p0, p1]);
 
     let vote = net.procs[1].propose(Reconfig::Join(rng.gen()))?;
     let resp = net.procs[0].handle_signed_vote(vote);
@@ -47,16 +49,16 @@ fn test_reject_vote_from_non_member() -> Result<()> {
 #[test]
 fn test_reject_join_if_actor_is_already_a_member() -> Result<()> {
     let mut rng = StdRng::from_seed([0u8; 32]);
-    let mut proc = State::<u8> {
+    let mut proc = MembershipState::<u8> {
         forced_reconfigs: vec![(
             0,
             BTreeSet::from_iter((0..1).map(|_| Reconfig::Join(rng.gen()))),
         )]
         .into_iter()
         .collect(),
-        ..State::random(&mut rng)
+        ..MembershipState::random(&mut rng)
     };
-    proc.elders = BTreeSet::from_iter([proc.public_key_share()]);
+    proc.state.elders = BTreeSet::from_iter([proc.public_key_share()]);
 
     let member = proc
         .members(proc.gen)?
@@ -73,16 +75,16 @@ fn test_reject_join_if_actor_is_already_a_member() -> Result<()> {
 #[test]
 fn test_reject_leave_if_actor_is_not_a_member() {
     let mut rng = StdRng::from_seed([0u8; 32]);
-    let mut proc = State::<u8> {
+    let mut proc = MembershipState::<u8> {
         forced_reconfigs: vec![(
             0,
             BTreeSet::from_iter((0..1).map(|_| Reconfig::Join(rng.gen()))),
         )]
         .into_iter()
         .collect(),
-        ..State::random(&mut rng)
+        ..MembershipState::random(&mut rng)
     };
-    proc.elders = BTreeSet::from_iter([proc.public_key_share()]);
+    proc.state.elders = BTreeSet::from_iter([proc.public_key_share()]);
 
     let resp = proc.propose(Reconfig::Leave(rng.gen()));
     assert!(matches!(resp, Err(Error::LeaveRequestForNonMember { .. })));
@@ -94,16 +96,16 @@ fn test_handle_vote_rejects_packet_from_previous_gen() -> Result<()> {
     let mut net = Net::with_procs(2, &mut rng);
     let a_0 = net.procs[0].public_key_share();
     let a_1 = net.procs[1].public_key_share();
-    let elders = BTreeSet::from_iter(net.procs.iter().map(State::public_key_share));
+    let elders = BTreeSet::from_iter(net.procs.iter().map(MembershipState::public_key_share));
     for proc in net.procs.iter_mut() {
-        proc.elders = elders.clone();
+        proc.state.elders = elders.clone();
     }
 
     let vote = net.procs[0].propose(Reconfig::Join(rng.gen()))?;
     let packets = net
         .procs
         .iter()
-        .map(State::public_key_share)
+        .map(MembershipState::public_key_share)
         .map(|dest| Packet {
             source: a_0,
             dest,
@@ -116,7 +118,7 @@ fn test_handle_vote_rejects_packet_from_previous_gen() -> Result<()> {
     let stale_packets = net
         .procs
         .iter()
-        .map(State::public_key_share)
+        .map(MembershipState::public_key_share)
         .map(|dest| Packet {
             source: a_1,
             dest,
@@ -125,7 +127,7 @@ fn test_handle_vote_rejects_packet_from_previous_gen() -> Result<()> {
         .collect::<Vec<_>>();
 
     net.procs[1].pending_gen = 0;
-    net.procs[1].votes = Default::default();
+    net.procs[1].state.votes = Default::default();
 
     assert_eq!(packets.len(), 2); // two members in the network
     assert_eq!(stale_packets.len(), 2);
@@ -150,14 +152,14 @@ fn test_handle_vote_rejects_packet_from_previous_gen() -> Result<()> {
 #[test]
 fn test_reject_votes_with_invalid_signatures() -> Result<()> {
     let mut rng = StdRng::from_seed([0u8; 32]);
-    let mut proc: State<u8> = State::random(&mut rng);
+    let mut proc: MembershipState<u8> = MembershipState::random(&mut rng);
     let ballot = Ballot::Propose(Reconfig::Join(rng.gen()));
     let gen = proc.gen + 1;
     let voter = rng.gen::<SecretKeyShare>().public_key_share();
     let bytes = bincode::serialize(&(&ballot, &gen))?;
     let sig = rng.gen::<SecretKeyShare>().sign(&bytes);
-    let vote = Vote { gen, ballot };
-    let resp = proc.handle_signed_vote(SignedVote { vote, voter, sig });
+    let vote = UnsignedVote { gen, ballot };
+    let resp = proc.handle_signed_vote(Vote { vote, voter, sig });
 
     assert!(resp.is_err());
     assert!(matches!(resp, Err(Error::InvalidElderSignature)));
@@ -171,9 +173,9 @@ fn test_split_vote() -> Result<()> {
     for nprocs in 1..7 {
         let mut net = Net::with_procs(nprocs, &mut rng);
 
-        let elders = BTreeSet::from_iter(net.procs.iter().map(State::public_key_share));
+        let elders = BTreeSet::from_iter(net.procs.iter().map(MembershipState::public_key_share));
         for proc in net.procs.iter_mut() {
-            proc.elders = elders.clone();
+            proc.state.elders = elders.clone();
         }
 
         for i in 0..net.procs.len() {
@@ -216,9 +218,9 @@ fn test_round_robin_split_vote() -> Result<()> {
     for nprocs in 1..7 {
         let mut net = Net::with_procs(nprocs, &mut rng);
 
-        let elders = BTreeSet::from_iter(net.procs.iter().map(State::public_key_share));
+        let elders = BTreeSet::from_iter(net.procs.iter().map(MembershipState::public_key_share));
         for proc in net.procs.iter_mut() {
-            proc.elders = elders.clone();
+            proc.state.elders = elders.clone();
         }
 
         for i in 0..net.procs.len() {
@@ -262,9 +264,9 @@ fn test_onboarding_across_many_generations() -> Result<()> {
     let p0 = net.procs[0].public_key_share();
     let p1 = net.procs[1].public_key_share();
 
-    let elders = BTreeSet::from_iter(net.procs.iter().map(State::public_key_share));
+    let elders = BTreeSet::from_iter(net.procs.iter().map(MembershipState::public_key_share));
     for proc in net.procs.iter_mut() {
-        proc.elders = elders.clone();
+        proc.state.elders = elders.clone();
     }
 
     let vote = net.procs[0].propose(Reconfig::Join(1)).unwrap();
@@ -310,9 +312,9 @@ fn test_simple_proposal() -> Result<()> {
     let mut rng = StdRng::from_seed([0u8; 32]);
     let mut net = Net::with_procs(3, &mut rng);
 
-    let elders = BTreeSet::from_iter(net.procs.iter().map(State::public_key_share));
+    let elders = BTreeSet::from_iter(net.procs.iter().map(MembershipState::public_key_share));
     for proc in net.procs.iter_mut() {
-        proc.elders = elders.clone();
+        proc.state.elders = elders.clone();
     }
 
     let p0 = net.procs[0].public_key_share();
@@ -408,9 +410,9 @@ fn test_interpreter_qc1() -> Result<()> {
     let mut net = Net::with_procs(2, &mut rng);
     let p0 = net.procs[0].public_key_share();
 
-    let elders = BTreeSet::from_iter(net.procs.iter().map(State::public_key_share));
+    let elders = BTreeSet::from_iter(net.procs.iter().map(MembershipState::public_key_share));
     for proc in net.procs.iter_mut() {
-        proc.elders = elders.clone();
+        proc.state.elders = elders.clone();
     }
 
     let reconfig = Reconfig::Join(1);
@@ -450,9 +452,9 @@ fn test_interpreter_qc2() -> Result<()> {
     let mut net = Net::with_procs(3, &mut rng);
     let p0 = net.procs[0].public_key_share();
 
-    let elders = BTreeSet::from_iter(net.procs.iter().map(State::public_key_share));
+    let elders = BTreeSet::from_iter(net.procs.iter().map(MembershipState::public_key_share));
     for proc in net.procs.iter_mut() {
-        proc.elders = elders.clone();
+        proc.state.elders = elders.clone();
     }
 
     let vote = net.procs[0].propose(Reconfig::Join(1))?;
@@ -470,7 +472,7 @@ fn test_interpreter_qc2() -> Result<()> {
     let expected_members = net.procs[0].members(net.procs[0].pending_gen)?;
     for p in net.procs.iter() {
         assert_eq!(p.gen, p.pending_gen);
-        assert_eq!(p.votes, Default::default());
+        assert_eq!(p.state.votes, Default::default());
         assert_eq!(p.members(p.gen)?, expected_members);
     }
 
@@ -482,9 +484,9 @@ fn test_interpreter_qc3() {
     let mut rng = StdRng::from_seed([0u8; 32]);
     let mut net = Net::with_procs(4, &mut rng);
 
-    let elders = BTreeSet::from_iter(net.procs.iter().map(State::public_key_share));
+    let elders = BTreeSet::from_iter(net.procs.iter().map(MembershipState::public_key_share));
     for proc in net.procs.iter_mut() {
-        proc.elders = elders.clone();
+        proc.state.elders = elders.clone();
     }
 
     let p0 = net.procs[0].public_key_share();
@@ -535,9 +537,9 @@ fn test_interpreter_qc3() {
 #[test]
 fn test_procs_refuse_to_propose_competing_votes() -> Result<()> {
     let rng = StdRng::from_seed([0u8; 32]);
-    let mut proc = State::random(rng);
+    let mut proc = MembershipState::random(rng);
     let proc_id = proc.public_key_share();
-    proc.elders = BTreeSet::from_iter([proc_id]);
+    proc.state.elders = BTreeSet::from_iter([proc_id]);
 
     proc.propose(Reconfig::Join(0))?;
 
@@ -548,10 +550,11 @@ fn test_procs_refuse_to_propose_competing_votes() -> Result<()> {
         Err(Error::ExistingVoteIncompatibleWithNewVote)
     ));
     assert!(!proc
+        .state
         .votes
         .get(&proc_id)
         .unwrap()
-        .supersedes(&proc.sign_vote(Vote {
+        .supersedes(&proc.sign_vote(UnsignedVote {
             ballot: Ballot::Propose(reconfig),
             gen: proc.gen,
         })?));
@@ -562,7 +565,7 @@ fn test_procs_refuse_to_propose_competing_votes() -> Result<()> {
 #[test]
 fn test_validate_reconfig_rejects_when_members_at_capacity() -> Result<()> {
     let rng = StdRng::from_seed([0u8; 32]);
-    let mut proc = State::random(rng);
+    let mut proc = MembershipState::random(rng);
 
     for m in 0..7 {
         proc.force_join(m);
@@ -585,18 +588,18 @@ fn test_bft_consensus_qc1() -> Result<()> {
         net.procs[5].public_key_share(),
     ]);
 
-    let elders = BTreeSet::from_iter(net.procs.iter().map(State::public_key_share));
+    let elders = BTreeSet::from_iter(net.procs.iter().map(MembershipState::public_key_share));
     for p in net.procs.iter_mut() {
-        p.elders = elders.clone();
+        p.state.elders = elders.clone();
     }
 
     // send a randomized packet
     let packet = Packet {
         source: net.procs[1].public_key_share(),
         dest: net.procs[0].public_key_share(),
-        vote: SignedVote {
+        vote: Vote {
             voter: net.procs[1].public_key_share(),
-            ..net.procs[1].sign_vote(Vote {
+            ..net.procs[1].sign_vote(UnsignedVote {
                 gen: 1,
                 ballot: Ballot::Propose(Reconfig::Join(240)),
             })?
@@ -606,9 +609,9 @@ fn test_bft_consensus_qc1() -> Result<()> {
     let packet = Packet {
         source: net.procs[1].public_key_share(),
         dest: net.procs[0].public_key_share(),
-        vote: SignedVote {
+        vote: Vote {
             voter: net.procs[1].public_key_share(),
-            ..net.procs[5].sign_vote(Vote {
+            ..net.procs[5].sign_vote(UnsignedVote {
                 gen: 0,
                 ballot: Ballot::Propose(Reconfig::Join(115)),
             })?
@@ -617,7 +620,7 @@ fn test_bft_consensus_qc1() -> Result<()> {
     net.enqueue_packets(vec![packet]);
 
     while let Err(e) = net.drain_queued_packets() {
-        println!("Error while draining: {e:?}");
+        println!("Error while draining: {:?}", e);
     }
     net.generate_msc("bft_consensus_qc1.msc")?;
 
@@ -630,7 +633,7 @@ fn test_bft_consensus_qc1() -> Result<()> {
     // BFT TERMINATION PROPERTY: all honest procs have decided ==>
     for p in honest_procs.iter() {
         assert_eq!(p.gen, p.pending_gen);
-        assert_eq!(p.votes, BTreeMap::default());
+        assert_eq!(p.state.votes, BTreeMap::default());
     }
 
     // BFT AGREEMENT PROPERTY: all honest procs have decided on the same values
@@ -662,9 +665,9 @@ fn prop_interpreter(n: u8, instructions: Vec<Instruction>, seed: u128) -> eyre::
 
     let mut net = Net::with_procs(n, &mut rng);
 
-    let elders = BTreeSet::from_iter(net.procs.iter().map(State::public_key_share));
+    let elders = BTreeSet::from_iter(net.procs.iter().map(MembershipState::public_key_share));
     for proc in net.procs.iter_mut() {
-        proc.elders = elders.clone();
+        proc.state.elders = elders.clone();
     }
 
     for instruction in instructions {
@@ -686,18 +689,21 @@ fn prop_interpreter(n: u8, instructions: Vec<Instruction>, seed: u128) -> eyre::
                         assert!(q.members(q.gen)?.contains(&p));
                     }
                     Err(Error::NotElder { .. }) => {
-                        assert!(!q.elders.contains(&q.public_key_share()));
+                        assert!(!q.state.elders.contains(&q.public_key_share()));
                     }
                     Err(Error::ExistingVoteIncompatibleWithNewVote) => {
                         // This proc has already committed to a vote this round
 
                         // This proc has already committed to a vote
-                        assert!(!q.votes.get(&q.public_key_share()).unwrap().supersedes(
-                            &q.sign_vote(Vote {
+                        assert!(!q
+                            .state
+                            .votes
+                            .get(&q.public_key_share())
+                            .unwrap()
+                            .supersedes(&q.sign_vote(UnsignedVote {
                                 ballot: Ballot::Propose(reconfig),
                                 gen: q.gen,
-                            })?
-                        ));
+                            })?));
                     }
                     Err(err) => {
                         // invalid request.
@@ -722,17 +728,22 @@ fn prop_interpreter(n: u8, instructions: Vec<Instruction>, seed: u128) -> eyre::
                         assert!(!q.members(q.gen)?.contains(&p));
                     }
                     Err(Error::NotElder { .. }) => {
-                        assert!(!q.elders.contains(&q.public_key_share()));
+                        assert!(!q.state.elders.contains(&q.public_key_share()));
                     }
                     Err(Error::ExistingVoteIncompatibleWithNewVote) => {
                         // This proc has already committed to a vote
-                        assert!(!q.votes.get(&q.public_key_share()).unwrap().supersedes(
-                            &q.sign_vote(Vote {
-                                ballot: Ballot::Propose(reconfig),
-                                gen: q.gen,
-                            })
+                        assert!(!q
+                            .state
+                            .votes
+                            .get(&q.public_key_share())
                             .unwrap()
-                        ))
+                            .supersedes(
+                                &q.sign_vote(UnsignedVote {
+                                    ballot: Ballot::Propose(reconfig),
+                                    gen: q.gen,
+                                })
+                                .unwrap()
+                            ))
                     }
                     Err(err) => {
                         // invalid request.
@@ -774,10 +785,10 @@ fn prop_interpreter(n: u8, instructions: Vec<Instruction>, seed: u128) -> eyre::
 
     // We should have no more pending votes.
     for p in net.procs.iter() {
-        assert_eq!(p.votes, Default::default());
+        assert_eq!(p.state.votes, Default::default());
     }
 
-    let mut procs_by_gen: BTreeMap<Generation, Vec<State<u8>>> = Default::default();
+    let mut procs_by_gen: BTreeMap<Generation, Vec<MembershipState<u8>>> = Default::default();
 
     for proc in net.procs {
         procs_by_gen.entry(proc.gen).or_default().push(proc);
@@ -835,7 +846,7 @@ fn prop_interpreter(n: u8, instructions: Vec<Instruction>, seed: u128) -> eyre::
     let proc_at_max_gen = procs_by_gen[max_gen].get(0).ok_or(Error::NoMembers)?;
     assert!(super_majority(
         procs_by_gen[max_gen].len(),
-        proc_at_max_gen.elders.len()
+        proc_at_max_gen.state.elders.len()
     ));
 
     Ok(TestResult::passed())
@@ -857,13 +868,13 @@ fn prop_validate_reconfig(
         return Ok(TestResult::discard());
     }
 
-    let mut proc = State::random(&mut rng);
+    let mut proc = MembershipState::random(&mut rng);
 
     for m in initial_members.iter().copied() {
         proc.force_join(m);
     }
 
-    proc.elders = iter::repeat_with(|| rng.gen::<SecretKeyShare>().public_key_share())
+    proc.state.elders = iter::repeat_with(|| rng.gen::<SecretKeyShare>().public_key_share())
         .take(num_elders as usize)
         .chain(iter::once(proc.public_key_share()))
         .collect();
@@ -933,9 +944,9 @@ fn prop_bft_consensus(
             .map(|idx| net.procs[idx as usize].public_key_share()),
     );
 
-    let elders = BTreeSet::from_iter(net.procs.iter().map(State::public_key_share));
+    let elders = BTreeSet::from_iter(net.procs.iter().map(MembershipState::public_key_share));
     for p in net.procs.iter_mut() {
-        p.elders = elders.clone();
+        p.state.elders = elders.clone();
     }
 
     let n_actions = rng.gen::<u8>() % 4;
@@ -962,7 +973,7 @@ fn prop_bft_consensus(
                     .procs
                     .iter_mut()
                     .filter(|p| !faulty.contains(&p.public_key_share())) // filter out faulty nodes
-                    .filter(|p| p.elders.contains(&p.public_key_share())) // filter out non-members
+                    .filter(|p| p.state.elders.contains(&p.public_key_share())) // filter out non-members
                     .filter(|p| p.gen != p.pending_gen) // filter out nodes who have already voted this round
                     .choose(&mut rng)
                 {
@@ -1001,7 +1012,7 @@ fn prop_bft_consensus(
     }
 
     while let Err(e) = net.drain_queued_packets() {
-        println!("Error while draining: {e:?}");
+        println!("Error while draining: {:?}", e);
     }
 
     let honest_procs = Vec::from_iter(
@@ -1012,7 +1023,7 @@ fn prop_bft_consensus(
 
     // BFT TERMINATION PROPERTY: all honest procs have decided ==>
     for p in honest_procs.iter() {
-        assert_eq!(p.votes, Default::default());
+        assert_eq!(p.state.votes, Default::default());
         assert_eq!(p.gen, p.pending_gen);
     }
